requirements.txt
beautifulsoup4==4.14.3
certifi==2026.4.22
charset-normalizer==3.4.7
idna==3.13
linkify-it-py==2.1.0
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mutagen==1.47.0
numpy==2.4.4
pillow==12.2.0
platformdirs==4.9.6
Pygments==2.20.0
python-mpv==1.0.8
python_mvp==0.1.0
RapidFuzz==3.14.5
requests==2.33.1
rich==15.0.0
soupsieve==2.8.3
syncedlyrics==1.0.1
textual==8.2.4
typing_extensions==4.15.0
uc-micro-py==2.0.0
urllib3==2.6.3
